refactor(gui): react-doctor cleanup for Startup/Debug/Storage/Usage
Split Startup/Debug sections and apply 0.9.1 fetch/loading hardening on Storage and Usage.
